Rename mm-builder modules in order to be more acurate with they actually do
