+++ lib/grainstore/index.js
@@ -5,5 +5,5 @@ module.exports = {
         return version;
     },
     MMLStore: require('./mml_store'),
-    MMLBuilder: require('./mml-builder/mml-builder-worker')
+    MMLBuilder: require('./mml-builder/mml-builder')
 };
