+++ lib/grainstore/mml-builder/mml-builder-child.js
@@ -1,6 +1,6 @@
 var _ = require('underscore');
 var debug = require('debug')('grainstore:mml-builder-child');
-var MMLBuilder = require('./mml_builder');
+var MMLBuilder = require('./mml-builder-inline');
 
 process.on('message', function(input) {
     debug('Child got MML Builder Context');
