+++ lib/grainstore/mml-builder/mml-builder-inline.js
@@ -2,8 +2,8 @@ var _      = require('underscore');
 var millstone = require('millstone');
 var StyleTrans = require('../style_trans');
 var carto = require('carto');
-var debug = require('debug')('grainstore:mml-builder');
-var debugXml = require('debug')('grainstore:mml-builder:xml');
+var debug = require('debug')('grainstore:mml-builder-inline');
+var debugXml = require('debug')('grainstore:mml-builder-inline:xml');
 
 // configure grainstore from optional args passed + defaults
 var grainstore_defaults = {
@@ -81,7 +81,7 @@ var supportedDatasourceTypes = {
 //       }
 //     }
 //
-function MMLBuilder(params, options) {
+function MMLBuilderInline(params, options) {
     this.params = params || {};
     // core variables
     var requiredParams = ['dbname', 'sql', 'style'];
@@ -133,9 +133,9 @@ function MMLBuilder(params, options) {
     }
 }
 
-module.exports = MMLBuilder;
+module.exports = MMLBuilderInline;
 
-MMLBuilder.prototype.set = function (property, value) {
+MMLBuilderInline.prototype.set = function (property, value) {
     var isOption = this.hasOwnProperty(property) && !_.isFunction(this[property]);
 
     if (!isOption) {
@@ -147,14 +147,14 @@ MMLBuilder.prototype.set = function (property, value) {
     return this; // allow chaining
 };
 
-MMLBuilder.prototype.toXML = function(callback) {
+MMLBuilderInline.prototype.toXML = function(callback) {
     var style = this.params.style;
     var style_version = this.params.style_version || this.default_style_version;
 
     this.render(style, style_version, callback);
 };
 
-MMLBuilder.prototype.render = function(style_in, version, callback){
+MMLBuilderInline.prototype.render = function(style_in, version, callback){
     var self = this;
 
     style_in = Array.isArray(style_in) ? style_in : [ style_in ];
@@ -228,7 +228,7 @@ MMLBuilder.prototype.render = function(style_in, version, callback){
     });
 };
 
-MMLBuilder.prototype.baseMML = function() {
+MMLBuilderInline.prototype.baseMML = function() {
     var tables = Array.isArray(this.params.sql) ? this.params.sql : [ this.params.sql ];
 
     var mml   = {};
@@ -297,7 +297,7 @@ MMLBuilder.prototype.baseMML = function() {
     return mml;
 };
 
-MMLBuilder.prototype.toMML = function(style_in){
+MMLBuilderInline.prototype.toMML = function(style_in){
 
     var base_mml = this.baseMML();
     base_mml.Stylesheet = [];
