+++ lib/grainstore/mml-builder/mml-builder.js
@@ -1,10 +1,10 @@
 var util = require('util');
-var MMLBuilder = require('./mml_builder');
+var MMLBuilderInline = require('./mml-builder-inline');
 
 var pool = require('generic-pool');
 var fork = require('child_process').fork;
-var debug = require('debug')('grainstore:mml-builder-worker');
-var debugXml = require('debug')('grainstore:mml-builder-worker:xml');
+var debug = require('debug')('grainstore:mml-builder');
+var debugXml = require('debug')('grainstore:mml-builder:xml');
 
 var childMMLBuilderPool = pool.Pool({
     name: 'mml-builder',
@@ -25,19 +25,19 @@ var childMMLBuilderPool = pool.Pool({
     log: false
 });
 
-function MMLBuilderWorker(params, options) {
-    MMLBuilder.call(this, params, options);
+function MMLBuilder(params, options) {
+    MMLBuilderInline.call(this, params, options);
 }
 
-util.inherits(MMLBuilderWorker, MMLBuilder);
+util.inherits(MMLBuilder, MMLBuilderInline);
 
-module.exports = MMLBuilderWorker;
+module.exports = MMLBuilder;
 
-MMLBuilderWorker.prototype.toXML = function(callback) {
+MMLBuilder.prototype.toXML = function(callback) {
     var self = this;
 
     if (!this.options.use_workers) {
-        return MMLBuilder.prototype.toXML.call(this, callback);
+        return MMLBuilderInline.prototype.toXML.call(this, callback);
     }
 
     debug('Acquiring child process');
