+++ lib/grainstore/mml_store.js
@@ -2,7 +2,7 @@ var fs = require('fs');
 var _ = require('underscore');
 var debug = require('debug')('grainstore:mmlstore');
 
-var MMLBuilderWorker = require('./mml-builder/mml-builder-worker');
+var MMLBuilder = require('./mml-builder/mml-builder');
 
 // @param optional_args
 //    optional configurations. valid elements:
@@ -21,7 +21,7 @@ function MMLStore(options) {
   // @param callback(err, payload) called on initialization
   me.mml_builder = function(params, overrideOptions) {
     var opts = _.extend({}, options, overrideOptions);
-    return new MMLBuilderWorker(params, opts);
+    return new MMLBuilder(params, opts);
   };
 
   /// API: Purge cache of localized resources for this store
